[chore] 상품 프로젝트 패키지 경로 변경
