--- src/main/java/com/pofol/main/product/admin/AdminController.java
@@ -1,7 +1,0 @@
-package com.pofol.main.product.admin;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class AdminController {
-}
