+++ src/main/java/com/pofol/main/product/controller/ProductController.java
@@ -1,23 +1,20 @@
 package com.pofol.main.product.controller;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.util.FileCopyUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.List;
 
 @Controller
 @Slf4j
-public class ProdController {
+public class ProductController {
 
     @GetMapping("/prodEnroll")
     public String prodEnrollPageGET() {
