+++ src/main/java/com/pofol/main/product/domain/OptionProductDto.java
@@ -1,0 +1,4 @@
+package com.pofol.main.product.domain;
+
+public class OptionProductDto {
+}
