--- src/main/java/com/pofol/main/product/list/domain/OptionProductDto.java
@@ -1,4 +1,0 @@
-package com.pofol.main.product.list.domain;
-
-public class OptionProductDto {
-}
