+++ src/main/resources/mybatis-config.xml
@@ -4,8 +4,7 @@
 		"http://mybatis.org/dtd/mybatis-3-config.dtd">
 <configuration>
 	<typeAliases>
-		<typeAlias alias="ProductDto" type="com.pofol.main.product.list.domain.ProductDto"/>
-		<typeAlias alias="EventGroupDto" type="com.pofol.main.product.list.domain.EventGroupDto"/>
 		<typeAlias alias="ProductDto" type="com.pofol.main.product.domain.ProductDto"/>
+		<typeAlias alias="EventGroupDto" type="com.pofol.main.product.domain.EventGroupDto"/>
 	</typeAliases>
 </configuration>
