+++ src/main/webapp/WEB-INF/web.xml
@@ -9,7 +9,7 @@
 		<param-name>contextConfigLocation</param-name>
 		<param-value>/WEB-INF/spring/root-context.xml</param-value>
 	</context-param>
-	
+
 	<!-- Creates the Spring Container shared by all Servlets and Filters -->
 	<listener>
 		<listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
@@ -28,12 +28,12 @@
 		<!-- 업로드 관련 설정 -->
 		<multipart-config>
 			<location>C:\\upload\\temp</location>
-			<max-file-size>20971520</max-file-size>					<!-- 1MB * 20 -->
-			<max-request-size>41943040</max-request-size>			<!-- 40MB -->
-			<file-size-threshold>20971520</file-size-threshold>		<!-- 20MB -->
+			<max-file-size>20971520</max-file-size>                <!-- 1MB * 20 -->
+			<max-request-size>41943040</max-request-size>        <!-- 40MB -->
+			<file-size-threshold>20971520</file-size-threshold>       <!-- 20MB -->
 		</multipart-config>
 	</servlet>
-		
+
 	<servlet-mapping>
 		<servlet-name>appServlet</servlet-name>
 		<url-pattern>/</url-pattern>
