+++ src/main/webapp/resources/css/product-css.css
@@ -152,7 +152,7 @@ html, body {
     content: "";
     position: absolute;
     display: block;
-    background: url(./kurly_new.svg) 0px 0px no-repeat;
+    background: url(/resources/icons/kurly_new.svg) 0px 0px no-repeat;
     width: 10px;
     height: 10px;
     top: 3px;
@@ -186,7 +186,7 @@ html, body {
     bottom: 3px;
     border: none;
     cursor: pointer;
-    background: url(./kurly_search.svg) 0px 0px no-repeat;
+    background: url(/resources/icons/kurly_search.svg) 0px 0px no-repeat;
 }
 #search > #search-bar > #search-right > button{
     width: 36px;
@@ -197,25 +197,25 @@ html, body {
     border: none;
 }
 #search > #search-bar > #search-right > button:nth-of-type(3) {
-    background: url(./kurly_cart.svg) 50% 50% no-repeat;
+    background: url(/resources/icons/kurly_cart.svg) 50% 50% no-repeat;
     right: 0;
 }
 #search > #search-bar > #search-right > button:nth-of-type(3):hover {
-    background: url(./kurly_cart_purple.svg) 50% 50% no-repeat;
+    background: url(/resources/icons/kurly_cart_purple.svg) 50% 50% no-repeat;
 }
 #search > #search-bar > #search-right > button:nth-of-type(2) {
-    background: url(./kurly_heart.svg) 50% 50% no-repeat;
+    background: url(/resources/icons/kurly_heart.svg) 50% 50% no-repeat;
     right: 58px;
 }
 #search > #search-bar > #search-right > button:nth-of-type(2):hover {
-    background: url(./kurly_heart_purple.svg) 50% 50% no-repeat;
+    background: url(/resources/icons/kurly_heart_purple.svg) 50% 50% no-repeat;
 }
 #search > #search-bar > #search-right > button:nth-of-type(1) {
-    background: url(./kurly_adressAdd.svg) 50% 50% no-repeat;
+    background: url(/resources/icons/kurly_adressAdd.svg) 50% 50% no-repeat;
     right: 116px;
 }
 #search > #search-bar > #search-right > button:nth-of-type(1):hover {
-    background: url(./kurly_adressAdd_purple.svg) 50% 50% no-repeat;
+    background: url(/resources/icons/kurly_adressAdd_purple.svg) 50% 50% no-repeat;
 }
 
 #nev {
@@ -235,10 +235,10 @@ html, body {
     width: 16px;
     height: 14px;
     margin-right: 14px;
-    background: url(./kurly_hamburger.svg) 0px 0px / 16px 14px no-repeat;
+    background: url(/resources/icons/kurly_hamburger.svg) 0px 0px / 16px 14px no-repeat;
 }
 #nev > #category:hover > span:nth-of-type(1) {
-    background: url(./kurly_hamburger_purple.svg) 0px 0px / 16px 14px no-repeat;
+    background: url(/resources/icons/kurly_hamburger_purple.svg) 0px 0px / 16px 14px no-repeat;
 }
 #nev > #category > span:nth-of-type(2) {
     font-size: 16px;
